% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptotic_information.R
\name{asymptotic_info}
\alias{asymptotic_info}
\alias{asymptotic_information_difference_means}
\alias{asymptotic_information_difference_proportions}
\alias{asymptotic_information_risk_difference}
\alias{asymptotic_information_relative_risk}
\title{Approximate information from a random samples of a given size for continuous and binary outcomes}
\usage{
asymptotic_information_difference_means(n_0, sigma_0, n_1, sigma_1)

asymptotic_information_difference_proportions(n_0, pi_0, n_1, pi_1)

asymptotic_information_risk_difference(...)

asymptotic_information_relative_risk(n_0, pi_0, n_1, pi_1)
}
\arguments{
\item{n_0}{Numeric vector containing the sample size in the control arm}

\item{sigma_0}{Variance of outcomes in the population of individuals
receiving the control intervention}

\item{n_1}{Numeric vector containing the sample size in the treatment arm}

\item{sigma_1}{Variance of outcomes in the population of individuals
receiving the active intervention}

\item{pi_0}{Probability of event in the population of individuals
receiving the control intervention}

\item{pi_1}{Probability of event in the population of individuals
receiving the control intervention}

\item{...}{Arguments passed to \link{asymptotic_information_difference_proportions}}
}
\value{
When all parameters are scalars, the result is a scalar, indicating
the approximate information. When multiple values are specified, a grid of
unique parameters are constructed, and the approximate information is
computed for each value of the parameters.
}
\description{
These functions provide an asymptotic approximation to the information
(i.e. precision, inverse of the variance) provided by two samples under
assumed values of nuisance parameters for continuous and binary outcomes.
}
\examples{
# When a single value is supplied for each parameter, a scalar is returned:
asymptotic_information_difference_means(
  n_0 = 50,
  sigma_0 = 5,
  n_1 = 50,
  sigma_1 = 5
)

# When multiple values are supplied for one or more parameters, the grid of
# parameters are created, and a data.frame is returned.
asymptotic_information_difference_means(
  n_0 = c(50, 75),
  sigma_0 = 5,
  n_1 = c(50, 75),
  sigma_1 = 5
)
}
\seealso{
\link{asymptotic_information_mann_whitney_fm} for ordinal outcomes.
}
