% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptotic_information_logrank.R
\name{asymptotic_information_logrank}
\alias{asymptotic_information_logrank}
\title{Approximate information from a random samples of a given size for log hazard ratio}
\usage{
asymptotic_information_logrank(allocation_ratio = 1, total_events)
}
\arguments{
\item{allocation_ratio}{Numeric scalar containing the allocation ratio of r
treatments to 1 control. Defaults to 1.}

\item{total_events}{Numeric vector containing the total number of events
observed across both treatment arms}
}
\value{
A \code{numeric} scalar or \code{data.frame} containing an
approximate information level for the values of the inputs.
}
\description{
Approximate information from a random samples of a given size for log hazard ratio
}
\examples{
asymptotic_information_logrank(
  allocation_ratio = 1,
  total_events = 90
)

asymptotic_information_logrank(
  allocation_ratio = 1,
  total_events = c(66, 90)
)

asymptotic_information_logrank(
  allocation_ratio = c(1, 2),
  total_events = c(66, 90)
)
}
