% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_control.R
\name{boot_control}
\alias{boot_control}
\alias{boot_control_testing}
\title{Parameters for using boot::boot() in Non-Sequential Analyses}
\usage{
boot_control(
  bootstrap_n = 10000,
  parallel = "no",
  ncpus = getOption("boot.ncpus", 1L),
  ...
)

boot_control_testing(bootstrap_n = 1000, ...)
}
\arguments{
\item{bootstrap_n}{A \code{numeric} scalar indicating the number of bootstrap
replicates to perform.}

\item{parallel}{A \code{character} scalar indicating whether and how to
use parallel computing: see \code{?\link[boot]{boot}}}

\item{ncpus}{A \code{numeric} scalar indicating the number of cores to use in
parallel computing: see \code{?\link[boot]{boot}}}

\item{...}{Other named parameters passed to \code{\link[boot]{boot}}}
}
\value{
A list containing the default parameters or supplied alternatives.
}
\description{
These functions are convenience functions for supplying parameters to
bootstrap-based analyses. \strong{NOTE}: sequential analyses require
computing the covariance of estimates using a bootstrap function that
preserves the covariance inherent in group sequential designs. This function
should not be used for group sequential designs.
}
\examples{
boot_control()
boot_control_testing()
}
\seealso{
\code{\link[parallel:clusterApply]{parallel::parSapplyLB()}} and \code{\link[parallel:clusterApply]{parallel::parSapply()}} for parallel
computing; \link{monitored_analysis} for conducting information monitored
analyses, which uses \link{estimate_information} and
\code{\link[=calculate_covariance]{calculate_covariance()}} for computing the covariance of estimates.
}
