% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariance.R
\name{calculate_covariance}
\alias{calculate_covariance}
\title{Compute bootstrap covariance of estimates}
\usage{
calculate_covariance(
  data,
  ids_by_analysis = NULL,
  bootstrap_ids = NULL,
  bootstrap_results = NULL,
  estimation_function,
  estimation_arguments,
  rng_seed,
  control = monitored_analysis_control()
)
}
\arguments{
\item{data}{A \code{data.frame} containing the data to be analyzed. A column named
\code{.id} indicates which observations correspond to each individual.}

\item{ids_by_analysis}{A \code{list} containing the IDs that were enrolled
before each analysis occurred}

\item{bootstrap_ids}{A \code{list} of IDs resampled at earlier analyses: this is
NULL in the first analysis}

\item{bootstrap_results}{A numeric matrix of estimates: the bootstrap
estimates for each analysis is stored in the corresponding column.}

\item{estimation_function}{A function whose arguments include a data.frame
named \code{data}}

\item{estimation_arguments}{A \code{list} of any additional arguments needed by
\code{estimation_function}}

\item{rng_seed}{A numeric scalar for seeding the L'Ecuyer pseudorandom number
generator}

\item{control}{A \code{list} of computing arguments. See
\link{monitored_analysis_control} for details about the default computing
arguments.}
}
\value{
A \code{list} containing the covariance matrix of estimates, a matrix
of bootstrap estimates, and a list of the
IDs of individuals in each bootstrap replicate at each analysis.
}
\description{
Compute bootstrap covariance of estimates
}
\examples{
# To be added
}
\seealso{
\code{\link[=calculate_covariance]{calculate_covariance()}} for computing the covariance matrix
of estimators across analyses, and \link{monitored_analysis_control} for
details about the default computing arguments.
}
