% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_estimate.R
\name{calculate_estimate}
\alias{calculate_estimate}
\title{Compute an estimate from a wrapper function}
\usage{
calculate_estimate(data, estimation_function, estimation_arguments)
}
\arguments{
\item{data}{A data.frame containing the data to be analyzed}

\item{estimation_function}{A function whose arguments include a data.frame
named \code{data}}

\item{estimation_arguments}{A list of any additional arguments needed by
\code{estimation_function}}
}
\value{
A numeric scalar containing the computed estimate
}
\description{
Compute an estimate from a wrapper function
}
\examples{
# To be added
}
\seealso{
\code{\link[=calculate_covariance]{calculate_covariance()}} for computing the covariance matrix
of estimators across analyses.
}
