% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colon_cancer}
\alias{colon_cancer}
\title{colon_cancer: Processed Moerton Colon Cancer Data}
\format{
\subsection{\code{colon_cancer}}{

A data frame with 929 rows and 15 columns:
\describe{
\item{.id}{Participant ID}
\item{arm}{Treatment Assignment}
\item{age}{Patient age}
\item{sex}{Patient sex}
\item{obstruction}{Colon obstruction by tumor}
\item{perforation}{Colon perforation by tumor}
\item{organ_adherence}{Organ adherence by tumor}
\item{positive_nodes}{Lymph nodes with detectable cancer}
\item{differentiation}{Differentiation of tumor}
\item{local_spread}{Extent of local spread of tumor}
\item{time_surgery_registration}{Time from surgery to registration}
\item{event_death}{Death on study (1 = yes, 0 = no)}
\item{event_recurrence}{Recurrence (1 = yes, 0 = no)}
\item{years_to_death}{Time to death/censoring in years}
\item{years_to_recurrence}{Time to recurrence/censoring in years}
}
}
}
\usage{
colon_cancer
}
\description{
Data from an adjuvant chemotherapy trial for colon cancer with variables
renamed and labeled, missing covariates imputed, and time scale converted
to years. For more information, see \code{?survival::colon}.
}
\keyword{datasets}
