% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bootstrap.R
\name{compute_bootstrap}
\alias{compute_bootstrap}
\alias{compute_bootstrap_serial}
\alias{compute_bootstrap_parallel}
\title{Compute bootstrap from a vector of IDs}
\usage{
compute_bootstrap_serial(data, ids, estimation_function, estimation_arguments)

compute_bootstrap_parallel(
  data,
  ids,
  estimation_function = estimation_function,
  estimation_arguments = estimation_arguments,
  n_cores = 1,
  use_load_balancing = FALSE,
  required_packages = character(0)
)
}
\arguments{
\item{data}{A \code{data.frame} containing the data to be analyzed. A column named
\code{.id} indicates which observations correspond to each individual.}

\item{ids}{A vector of IDs to resample from \code{data}}

\item{estimation_function}{A function whose arguments include a data.frame
named \code{data}}

\item{estimation_arguments}{A \code{list} of any additional arguments needed by
\code{estimation_function}}

\item{n_cores}{Scalar number of cores to use.}

\item{use_load_balancing}{Logical scalar: Should load balancing be used?}

\item{required_packages}{A character vector of required packages: extracted
from the control argument to \code{\link[=calculate_covariance]{calculate_covariance()}}}
}
\value{
A vector containing the results of \code{estimation_function} evaluated on
each bootstrap replicate of \code{data} contained in the columns of \code{ids}
}
\description{
Bootstrapping is done by resampling all rows corresponding to a resampled ID
to work with both long and wide format data without reshaping.
}
\examples{
# to be added

}
\seealso{
\code{\link[=calculate_covariance]{calculate_covariance()}} for computing the covariance matrix
of estimators across analyses, \link{monitored_analysis_control} for
details about the default computing arguments.
}
