% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_one_sided_gsd.R
\name{correct_one_sided_gsd}
\alias{correct_one_sided_gsd}
\title{Correct one-sided trialDesignGroupSequential object}
\usage{
correct_one_sided_gsd(trial_design, higher_better = TRUE)
}
\arguments{
\item{trial_design}{A \code{trialDesignGroupSequential} which was called with the
argument \code{sided = 1} (i.e. one-sided alternative).}

\item{higher_better}{Logical scalar: Should the alternative be
theta > theta_0 (TRUE) or theta < theta_0 (FALSE) or}
}
\value{
A \code{trialDesignGroupSequential} object with the efficacy and futility
boundaries corrected and a logical scalar \code{directionUpper} indicating whether
the alternative is greater than the null value (TRUE) or less than the null
value (FALSE).
}
\description{
Currently in \code{\link[rpact:getDesignGroupSequential]{rpact::getDesignGroupSequential()}}, when the test is one-sided,
there is no way to specify the direction of the null and alternative
hypotheses. This is an
\href{https://github.com/rpact-com/rpact/issues/26}{open issue in rpact.}. This
correction is only necessary for using \code{rpact} and \code{impart} together or when
exporting the correct boundaries from a one-sided
\code{trialDesignGroupSequential} object when the alternative is less than the
null value.
}
\examples{
# One sided test: Higher values = Better
trial_design_one_sided_higher <-
  rpact::getDesignGroupSequential(
    alpha = 0.05,
    beta = 1 - 0.80,
    sided = 1,
    informationRates = c(0.50, 0.75, 1),
    typeOfDesign = "asOF",
    typeBetaSpending = "bsOF",
    bindingFutility = FALSE
  )

# Bounds assumed are for $H_{A}: \theta > \theta_{0}$
trial_design_one_sided_higher$criticalValues
trial_design_one_sided_higher$futilityBounds

# One sided test: Lower values = Better
trial_design_one_sided_lower <-
  correct_one_sided_gsd(
    trial_design = trial_design_one_sided_higher,
    higher_better = FALSE
  )

# Bounds for $H_{A}: \theta < \theta_{0}$
trial_design_one_sided_lower$criticalValues
trial_design_one_sided_lower$futilityBounds
}
