% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_outcomes.R
\name{count_outcomes}
\alias{count_outcomes}
\title{Count outcome events from prepared study data}
\usage{
count_outcomes(prepared_data, study_time = NULL)
}
\arguments{
\item{prepared_data}{A prepeared dataset: see
\link{prepare_monitored_study_data}}

\item{study_time}{A \code{numeric} scalar indicating the study time at which
the data should be reconstructed and events should be counted}
}
\value{
A \code{data.frame} in long format, with one row per individual per
event (i.e. randomization, outcomes).
}
\description{
This function uses the timing of events, such as randomization times and
times of outcomes, to reconstruct the data available at an earlier point in
the study. This can be used for monitoring the rate of information accrual
over the course of a study. This function returns the count of outcomes by
study time.
}
\examples{
# To be added
}
