% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_outcomes_at_time_t.R
\name{count_outcomes_at_time_t}
\alias{count_outcomes_at_time_t}
\title{Reconstruct the count of events at a given study time during a study}
\usage{
count_outcomes_at_time_t(prepared_data, study_times)
}
\arguments{
\item{prepared_data}{A prepeared dataset: see
\link{prepare_monitored_study_data}}

\item{study_times}{A \code{numeric} vector of study times at which to count
events.}
}
\value{
A \code{data.frame}
}
\description{
This function uses the timing of events, such as randomization times and
times of outcomes, to reconstruct the data available at an earlier point in
the study. This can be used for monitoring the rate of information accrual
over the course of a study. This returns the number of events at a particular
point during the study.
}
\examples{
# To be added
}
