% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_at_time_t.R
\name{data_at_time_t}
\alias{data_at_time_t}
\title{Reconstruct data available at an earlier point in a study}
\usage{
data_at_time_t(prepared_data, study_time)
}
\arguments{
\item{prepared_data}{A prepeared dataset: see
\link{prepare_monitored_study_data}}

\item{study_time}{A \code{numeric} scalar indicating the study time at which
the data should be reconstructed}
}
\value{
A list
}
\description{
Reconstruct data available at an earlier point in a study
}
\examples{
# To be added
}
