% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_wls.R
\name{dr_wls}
\alias{dr_wls}
\alias{dr_wls_fit}
\title{Doubly Robust Estimation Using Joffe's Doubly Robust Weighted Least Squares (DR-WLS) Estimator}
\usage{
dr_wls(
  data,
  outcome_formula,
  outcome_family,
  treatment_formula,
  missing_formula,
  missing_family = quasibinomial,
  estimand = c("difference", "ratio", "oddsratio")[1],
  se_method = c("none", "bootstrap")[1],
  alpha = 0.05,
  bootstrap_parameters = bootstrap_parameters(),
  bootstrap_ci_method = "bca",
  verbose = FALSE
)

dr_wls_fit(
  data,
  outcome_formula,
  outcome_family,
  treatment_column,
  missing_formula,
  missing_family = quasibinomial,
  missing_column,
  predictors,
  estimand = c("difference", "ratio", "oddsratio")[1],
  verbose = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{outcome_formula}{A \code{formula} for the outcome model - passed to
\code{\link[stats]{glm}}}

\item{outcome_family}{A \code{family} for the outcome model - passed to
\code{\link[stats]{glm}}}

\item{treatment_formula}{A \code{formula}: the left-hand side should be a
binary variable indicating assignment to treatment, and the right-hand side
should be an intercept (e.g. \code{tx ~ 1})}

\item{missing_formula}{A \code{formula} for the censoring model - passed to
\code{\link[stats]{glm}}}

\item{missing_family}{A \code{family} for the censoring model - passed to
\code{\link[stats]{glm}}}

\item{estimand}{A \code{character} scalar specifying the estimate to compute,
such as \code{"difference"}, \code{"ratio"}, \code{"oddsratio"}}

\item{se_method}{A \code{character} scalar indicating how standard errors
and confidence intervals should be calculated: "none" only computes the
estimate, the "sandwich" method use asymptotic variance estimates.}

\item{alpha}{A \code{numeric} scalar, indicating the Type I error rate
(1 - Confidence Level)}

\item{bootstrap_parameters}{A \code{list} of arguments to be passed to
\code{\link[boot]{boot}} for computing the bootstrap.}

\item{bootstrap_ci_method}{A \code{character} vector, indicating the
method(s) used for computing confidence intervals. See
\code{\link[boot]{boot.ci}} for details.}

\item{verbose}{A \code{logical} scalar: should just the estimate be returned
(\code{FALSE}), or should a \code{list} be returned, containing intermediate
results (\code{TRUE})? Defaults to \code{FALSE}.}

\item{treatment_column}{A \code{character} scalar containing the name of
the treatment assignment variable}

\item{missing_column}{A \code{character} scalar containing the name of
the outcome missingness indicator variable}

\item{predictors}{A \code{character} vector containing the name of
covariates in either the outcome or censoring models.}
}
\value{
A \code{numeric} scalar, a \code{data.frame}, or a \code{list},
depending on the options \code{se_method} and \code{verbose}.
}
\description{
Doubly robust estimators provide a consistent estimate of their target if
either the outcome model or censoring model are correctly specified.
}
\examples{
# To be added
}
