% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_information.R
\name{estimate_information}
\alias{estimate_information}
\title{Estimate the observed information level}
\usage{
estimate_information(
  data,
  monitored_design,
  estimation_function,
  estimation_arguments,
  orthogonalize = NULL,
  rng_seed,
  return_results = FALSE,
  control = monitored_analysis_control()
)
}
\arguments{
\item{data}{A data.frame containing the data to be analyzed. A column named
\code{.id} indicates which observations correspond to each individual.}

\item{monitored_design}{An object of class \code{monitored_design} created
using \code{\link[=initialize_monitored_design]{initialize_monitored_design()}}}

\item{estimation_function}{A function whose arguments include a data.frame
named \code{data}}

\item{estimation_arguments}{A list of any additional arguments needed by
\code{estimation_function}}

\item{orthogonalize}{Logical scalar: Should estimates, their covariance,
and the resulting test statistics be orthogonalized?}

\item{rng_seed}{Numeric scalar containing the L'Ecuyer pseudorandom
number generator seed}

\item{return_results}{Logical scalar: Should estimates be returned, or only
the covariance and information levels?}

\item{control}{A list containing the control arguments for computation,
typically created with \code{\link[=monitored_analysis_control]{monitored_analysis_control()}}}
}
\value{
A list containing the following elements:
}
\description{
Estimate the observed information level
}
\examples{
# To be added
}
\seealso{
\code{\link[=calculate_covariance]{calculate_covariance()}} for computing the covariance matrix
of estimators across analyses, \link{monitored_analysis_control} for
details about the default computing arguments, \link{monitored_analysis}
for conducting analyses at pre-specified information levels
}
