% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_design.R
\name{hr_events}
\alias{hr_events}
\alias{hr_power}
\alias{hr_alpha}
\alias{hr_minimal}
\alias{hr_design}
\title{Design Calculations for a Hazard Ratio Estimand}
\usage{
hr_events(hazard_ratio, power = 0.8, alpha = 0.05, test_sides = 2, ratio = 1)

hr_power(events, hazard_ratio, alpha = 0.05, test_sides = 2, ratio = 1)

hr_alpha(events, hazard_ratio, power = 0.8, test_sides = 2, ratio = 1)

hr_minimal(events, power = 0.8, alpha = 0.05, test_sides = 2, ratio = 1)

hr_design(
  events = NULL,
  hazard_ratio = NULL,
  power = NULL,
  alpha = NULL,
  test_sides = 2,
  ratio = 1
)
}
\arguments{
\item{hazard_ratio}{A \code{numeric} vector: the minimum hazard ratio of interest}

\item{power}{A \code{numeric} vector: the statistical power of interest}

\item{alpha}{A \code{numeric} vector: Type I Error probability}

\item{test_sides}{A \code{numeric} vector: Number of sides for test (1 or 2)}

\item{ratio}{A \code{numeric} vector: the allocation ratio of treatment to
control, i.e. for an \eqn{r:1} trial, \code{ratio} = \eqn{r}.}

\item{events}{A \code{numeric} vector: the number of observed events}
}
\value{
a \code{numeric} vector or \code{data.frame}, depending on whether
arguments are specified as scalars or vectors.
}
\description{
This is a function for designing a study with a time-to-event outcome whose
inferential target is the hazard ratio. \code{hr_design()} will calculate
the minimum number of events, minimum hazard ratio, power, or Type I error
when the other arguments are specified.
}
\details{
The function \code{hr_design} takes in user input, and depending on the
arguments supplied, calls the appropriate function depending on whether the
number of pooled events observed (\code{hr_events}), power (\code{hr_power}),
Type I Error (\code{hr_alpha}), or relative risk (\code{hr_minimal}) is left
unspecified.
}
\examples{
# Number of Events
hr_design(
  hazard_ratio = 0.75,
  power = 0.80,
  alpha = 0.05,
  test_sides = 2
)

# Power
hr_design(
  events = 400,
  hazard_ratio = 0.75,
  alpha = 0.05,
  test_sides = 2
)

# Type I Error
hr_design(
  events = 400,
  hazard_ratio = 0.75,
  power = 0.80,
  test_sides = 2
)

# Hazard Ratio
hr_design(
  events = 400,
  power = 0.80,
  alpha = 0.05,
  test_sides = 2
)
}
\references{
{
Schoenfeld, DA. 1983. “Sample-Size Formula for the Proportional-Hazards
Regression Model.” Biometrics 39 (2): 499.
\url{https://doi.org/10.2307/2531021}.
}
}
\seealso{
\code{\link[stats:power.t.test]{stats::power.t.test()}} for design calculations for a difference in
means, \code{\link[stats:power.prop.test]{stats::power.prop.test()}} for design calculations for a difference in
proportions, \code{\link[=rr_design]{rr_design()}} for design calculations for a relative
risk estimand.
}
