% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impart-package.R
\docType{package}
\name{impart-package}
\alias{impart}
\alias{impart-package}
\title{impart: Information Monitoring for Precision Adaptive Randomized Trials (IMPART)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
