% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_covariates_mean_mode.R
\name{impute_covariates_mean_mode}
\alias{impute_covariates_mean_mode}
\title{Impute missing baseline covariates using mean/mode imputation}
\usage{
impute_covariates_mean_mode(
  data,
  baseline_covariates,
  binary_categorical = TRUE
)
}
\arguments{
\item{data}{a \code{data.frame} to impute}

\item{baseline_covariates}{A \code{character} \code{vector} containing the
names in \code{data} to impute.}

\item{binary_categorical}{A \code{logical} scalar indicating whether binary
variables (logicals, numeric columns containing only
\code{0}/\code{1}/\code{NA}) should be imputed as a categorical variable
(using the mode) or a continuous variable (using the mean). The latter
preserves the type of the column in output.}
}
\value{
A \code{data.frame} after imputing the columns named in
\code{baseline_covariates}
}
\description{
This function is for impute missing values in baseline covariates. For
\code{numeric} columns, the mean value of the observed values is imputed.
For \code{factor} columns, the mode value of the observed values is imputed.
}
\examples{
set.seed(12345)
mtcars_missing <- mtcars
mtcars_missing$cyl <- factor(mtcars_missing$cyl)
mtcars_missing$vs <- factor(mtcars_missing$vs)
mtcars_missing$am <- factor(mtcars_missing$am)
mtcars_missing$gear <- factor(mtcars_missing$gear)
for(i in 1:ncol(mtcars_missing))
  mtcars_missing[sample(x = 1:nrow(mtcars_missing), size = 3), i] <- NA

impute_covariates_mean_mode(
  data = mtcars_missing,
  baseline_covariates = names(mtcars_missing)
)
}
\references{
{
Benkeser, D, Díaz, I, Luedtke, A, Segal, J, Scharfstein, D, and Rosenblum,
M. 2020. "Improving Precision and Power in Randomized Trials for COVID-19
Treatments Using Covariate Adjustment, for Binary, Ordinal, and Time-to-Event
Outcomes." \emph{Biometrics} 77 (4): 1467–81.
\url{https://doi.org/10.1111/biom.13377}.
}
}
