% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptotic_information_mann_whitney.R
\name{mann_whitney}
\alias{mann_whitney}
\alias{asymptotic_information_mann_whitney_fm}
\alias{mw_from_pmfs}
\title{Compute Approximate Information from Sample Size: Mann-Whitney Estimand}
\usage{
asymptotic_information_mann_whitney_fm(
  n_0,
  n_1,
  mw = NULL,
  pmf_1 = NULL,
  pmf_0 = NULL,
  adjust = TRUE
)

mw_from_pmfs(pmf_0, pmf_1, reverse_scale = FALSE)
}
\arguments{
\item{n_0}{A \code{numeric} vector containing the sample size in the control
arm.}

\item{n_1}{A \code{numeric} vector containing the sample size in the
treatment arm.}

\item{mw}{A \code{numeric} vector containing the Mann-Whitney estimand.}

\item{pmf_1}{A \code{numeric} vector or matrix of row vectors, each
containing the probability mass function of outcomes in the population of
individuals receiving the active intervention.}

\item{pmf_0}{A \code{numeric} vector or matrix of row vectors, each
containing the probability mass function of outcomes in the population of
individuals receiving the control intervention.}

\item{adjust}{A \code{logical} scalar, indicating whether an adjustment for
ties should be performed. \emph{Note}: this can only be computed when
\code{pmf_0} and \code{pmf_1} are supplied.}

\item{reverse_scale}{A \code{logical} scalar: should the scales be reversed
when calculating the Mann-Whitney estimand? This may be useful when lower
categories indicate a preferable outcome.}
}
\value{
When all parameters are scalars, the result is a scalar, indicating
the approximate information. When multiple values are specified, a grid of
unique parameters are constructed, and the approximate information is
computed for each value of the parameters.
}
\description{
This function provides an asymptotic approximation to the information
(i.e. precision, inverse of the variance) provided by two samples under
assumed values of nuisance parameters for an ordinal outcome, analyzed using
the Mann-Whitney estimand. This can be obtained by specifying the value of
the Mann-Whitney estimand or the probability mass functions of outcomes in
each treatment arm. These functions may be useful in pre-trial
planning to determine when analyses may occur under different assumptions
about the nuisance parameters involved.
}
\details{
The amount of information in a sample of size \eqn{N} depends on
\emph{nuisance parameters}, such as the variance of continuous outcomes,
the risk of binary and time-to-event outcomes, rates of misssing data, and
the correlation between covariates and the outcomes of interest.

In studies with a fixed sample size, this sample size is chosen based on
assumptions about these nuisance parameters, which are incorporated into the
\emph{effect size}. The sample size is chosen to give power \eqn{(1 - \beta)}
while maintaining a type I error rate of \eqn{(\alpha)} under some assumed
effect size. Inaccurate estimates of nuisance parameters can lead to
over-powered or under-powered studies.

In an information-monitored design, investigators choose an estimand of
interest, such as the difference in means or proportions, that is free from
nuisance parameters. A trial is designed to identify some minimum important
difference \eqn{\delta_{min}} in the estimand with power \eqn{(1 - \beta)}
while maintaining a type I error rate of \eqn{(\alpha)}. Data is collected
until the precision of the estimate (i.e. the reciprocal of its variance)
reaches a pre-specified threshold \eqn{\mathcal{I}}:

\deqn{\mathcal{I} = \left(\frac{Z_{\alpha/s} + Z_{\beta}}{\delta_{min}}\right)^2
\approx \frac{1}{Var(\hat{\delta})} =
\frac{1}{\left(SE(\hat{\delta})\right)^2}}

The sample size required to reach the information target \eqn{\mathcal{I}}
depends on nuisance parameters mentioned above.

These functions allow a user to determine an approximate amount of
information contained in a sample of size \eqn{N} based on some assumptions
about the nuisance parameters.
}
\examples{
# When a single value is supplied for each parameter, a scalar is returned:
asymptotic_information_mann_whitney_fm(
    n_0 = 100,
    n_1 = 100,
    mw = 0.75,
    adjust = FALSE
  )

# When multiple values are supplied for one or more parameters, the grid of
# parameters are created, and a data.frame is returned.
asymptotic_information_mann_whitney_fm(
  n_0 = c(100, 150),
  n_1 = c(100, 150),
  mw = 0.75,
  adjust = FALSE
)


# Specifying PMFs - With and Without Tie Adjustment
asymptotic_information_mann_whitney_fm(
  n_0 = 100,
  n_1 = 100,
  pmf_0 = c(0.2, 0.2, 0.6),
  pmf_1 = c(0.1, 0.1, 0.8),
  adjust = TRUE
)

# Specifying Multiple PMFs
asymptotic_information_mann_whitney_fm(
  n_0 = 100,
  n_1 = 100,
  pmf_0 =
    rbind(
      c(0.2, 0.2, 0.6),
      c(0.3, 0.1, 0.6)
    ),
  pmf_1 =
    rbind(
      c(0.1, 0.1, 0.8),
      c(0.05, 0.05, 0.9)
      ),
  adjust = TRUE
)
}
\references{
{
Fay, MP and Malinovsky, Y. 2018. "Confidence Intervals of the Mann-Whitney
Parameter That Are Compatible with the Wilcoxon-Mann-Whitney Test."
\emph{Statistics in Medicine} 37 (27): 3991–4006.
\url{https://doi.org/10.1002/sim.7890}.

Zhao, YD, Rahardja D, and Qu Y. 2007. "Sample Size Calculation for the
Wilcoxon–Mann–Whitney Test Adjusting for Ties." \emph{Statistics in Medicine}
27 (3): 462–68. https://doi.org/10.1002/sim.2912.

Benkeser, D, Díaz, I, Luedtke, A, Segal, J, Scharfstein, D, and Rosenblum,
M. 2020. "Improving Precision and Power in Randomized Trials for COVID-19
Treatments Using Covariate Adjustment, for Binary, Ordinal, and Time-to-Event
Outcomes." \emph{Biometrics} 77 (4): 1467–81.
\url{https://doi.org/10.1111/biom.13377}.
}
}
\seealso{
\link{asymptotic_information_difference_means} for the
information on the difference in means,
\link{asymptotic_information_difference_proportions} for the information
on a difference in proportions (i.e. a risk difference),
\link{asymptotic_information_relative_risk} for the information on the
relative risk (i.e. risk ratio), and \link{asymptotic_information_logrank}
for information on the log hazard ratio estimand.
}
