% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitored_analysis.R
\name{monitored_analysis}
\alias{monitored_analysis}
\title{Perform pre-specified analyses for an interim monitored trial design}
\usage{
monitored_analysis(
  data,
  monitored_design = NULL,
  estimation_function,
  estimation_arguments,
  correction_function = NULL,
  control = monitored_analysis_control(),
  ...
)
}
\arguments{
\item{data}{A data.frame containing the data to be analyzed. A column named
\code{.id} indicates which observations correspond to each individual.}

\item{monitored_design}{An object of class \code{monitored_design} created
using \code{\link[=initialize_monitored_design]{initialize_monitored_design()}}}

\item{estimation_function}{A function whose arguments include a data.frame
named \code{data}}

\item{estimation_arguments}{A list of any additional arguments needed by
\code{estimation_function}}

\item{correction_function}{A function which takes the arguments supplied to
\code{estimation_function} and returns a numeric scalar which performs a
small sample correction to the variance estimate}

\item{control}{A list containing the control arguments for computation,
typically created with \code{\link[=monitored_analysis_control]{monitored_analysis_control()}}}

\item{...}{Additional arguments, currently ignored}
}
\value{
An object of class \code{monitored_design}
}
\description{
Perform pre-specified analyses for an interim monitored trial design
}
\examples{
# To be added
}
