% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitored_design_checks.R
\name{monitored_design_checks}
\alias{monitored_design_checks}
\title{Check consistency of a \code{monitored_design} object}
\usage{
monitored_design_checks(
  data,
  monitored_design = NULL,
  estimation_function,
  estimation_arguments,
  correction_function = NULL
)
}
\arguments{
\item{data}{A data.frame containing the data to be analyzed. A column named
\code{.id} indicates which observations correspond to each individual.}

\item{monitored_design}{An object of class \code{monitored_design} created
using \code{\link[=initialize_monitored_design]{initialize_monitored_design()}}}

\item{estimation_function}{A function whose arguments include a data.frame
named \code{data}}

\item{estimation_arguments}{A list of any additional arguments needed by
\code{estimation_function}}

\item{correction_function}{A function which takes the arguments supplied to
\code{estimation_function} and returns a numeric scalar which performs a
small sample correction to the variance estimate}
}
\value{
No result is returned.
}
\description{
Check consistency of a \code{monitored_design} object
}
\examples{
# To be added
}
