% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_from_pmfs.R
\name{mw_from_pmfs}
\alias{mw_from_pmfs}
\title{Title Compute the Mann-Whitney estimand from Probability Mass Functions}
\usage{
mw_from_pmfs(pmf_0, pmf_1, reverse_scale = FALSE)
}
\arguments{
\item{pmf_0}{A vector containing the PMF of outcomes in the control
population, i.e. Pr(Y = 1 | A = 0), Pr(Y = 2 | A = 0), ...}

\item{pmf_1}{A vector containing the PMF of outcomes in the treated
population, i.e. Pr(Y = 1 | A = 1), Pr(Y = 2 | A = 1), ...}

\item{reverse_scale}{(Scalar: Logical) Should the scales be reversed when
calculating the Mann-Whitney estimand? This may be useful when lower
categories indicate a preferable outcome.}
}
\value{
A scalar containing the Mann-Whitney Estimand
}
\description{
The Mann-Whitney estimand is the tie-adjusted probability of a randomly
selected individual from the population of treated individuals having a
an outcome with a higher outcome category level than a randomly selected
individual from the population of control individuals.
}
\examples{
mw_from_pmfs(
  pmf_0 = c(0.1, 0.1, 0.2, 0.6),
  pmf_1 = c(0.2, 0.2, 0.3, 0.3)
)

mw_from_pmfs(
  pmf_0 = c(0.1, 0.1, 0.2, 0.6),
  pmf_1 = c(0.2, 0.2, 0.3, 0.3),
  reverse_scale = TRUE
)
}
