% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalize_estimates.R
\name{orthogonalize_estimates}
\alias{orthogonalize_estimates}
\title{Orthogonalize estimates and covariance matrix}
\usage{
orthogonalize_estimates(estimates, covariance)
}
\arguments{
\item{estimates}{A numeric vector of estimates}

\item{covariance}{A numeric matrix whose number of rows and columns is equal
to the length of \code{estimates}}
}
\value{
a \code{list} containing the orthogonalized estimates and covariance
matrix.
}
\description{
Orthogonalize estimates and covariance matrix
}
\examples{
# To be added
}
