% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcome_counts.R
\name{plot_outcome_counts}
\alias{plot_outcome_counts}
\title{Title}
\usage{
plot_outcome_counts(
  prepared_data,
  study_time = NULL,
  type = "tc",
  count_increment = 10,
  time_increment = 30,
  color_palette = NULL
)
}
\arguments{
\item{prepared_data}{A prepeared dataset: see
\link{prepare_monitored_study_data}}

\item{study_time}{A \code{numeric} scalar indicating the study time at which
the data should be reconstructed and events should be counted}

\item{type}{A \code{character} scalar containing "t" for the total number of
observations, "c" for the number of complete observations, and "e" for the
number of events (for time-to-event or binary outcomes)}

\item{count_increment}{Plots have horizontal lines at regular increments to
assist in reading of counts. \code{count_increment} is a \code{numeric}
scalar indicating the increment between horizontal lines on the count axis.}

\item{time_increment}{Plots have horizontal lines at regular increments to
assist in reading of counts. \code{time_increment} is a \code{numeric}
scalar indicating the increment between vertical lines on the time axis.}

\item{color_palette}{A vector of colors for each event.}
}
\description{
Title
}
\examples{
# To be added
}
