% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_monitored_study_data.R
\name{prepare_monitored_study_data}
\alias{prepare_monitored_study_data}
\title{Prepare monitored study data for monitoring and analysis}
\usage{
prepare_monitored_study_data(
  data,
  study_time,
  id_variable,
  covariates_variables,
  enrollment_time_variable,
  treatment_variable,
  outcome_variables,
  outcome_time_variables,
  observe_missing_times = NULL,
  outcomes_sequential = TRUE,
  time_to_event = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame} containing the baseline covariates, treatment
assignment, randomization time, observed outcomes, and times of outcome
assessments.}

\item{study_time}{A \code{numeric} scalar indicating the time from study
initiation to the export of the data (or latest observed event, such as
randomization of participants or outcome ascertainment).}

\item{id_variable}{A \code{character} scalar containing the column name of
the study participant identifier.}

\item{covariates_variables}{A \code{character} vector containing the column
names of the covariates and outcomes observed prior to randomization.}

\item{enrollment_time_variable}{A \code{character} scalar containing the
column name of the enrollment/randomization time variable.}

\item{treatment_variable}{A \code{character} scalar containing the
column name of the binary treatment assignment variable.}

\item{outcome_variables}{A \code{character} vector containing the column
names of the outcomes observed after randomization.}

\item{outcome_time_variables}{A \code{character} vector containing the column
names of the observation times of outcomes observed after randomization.}

\item{observe_missing_times}{A \code{numeric} a vector containing the time
from randomization to the closing of each study window when
\code{time_to_event == FALSE}.}

\item{outcomes_sequential}{A \code{logical} scalar: are outcomes sequential
in nature (e.g. a sequence of study visits) or not (e.g. times from
randomization to different events)?}

\item{time_to_event}{A \code{logical} scalar: are outcomes times from
randomization to events (TRUE) or a continuous, binary, or ordinal outcome
(FALSE)?}
}
\value{
a \code{list} containing the prepared data, the original data,
and the variables specified above.
}
\description{
Prepare monitored study data for monitoring and analysis
}
\examples{
# To be added
}
