% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_information_sequential.R
\name{required_information_sequential}
\alias{required_information_sequential}
\title{Adjust Information for Interim Monitoring}
\usage{
required_information_sequential(information_fixed, trial_design)
}
\arguments{
\item{information_fixed}{A numeric scalar containing the information level
for an information monitored design with a single efficacy analysis, created
using \code{\link[=required_information_fixed_n]{required_information_fixed_n()}}}

\item{trial_design}{An object of type \code{trialDesignGroupSequential}
created by \code{\link[rpact:getDesignGroupSequential]{rpact::getDesignGroupSequential()}} containing a group sequential
design specification}
}
\value{
A scalar information level which accounts for the pre-planned
analyses specified in \code{gsd_design_specification}
}
\description{
Group Sequential Design methods can control the Familywise Type I Error
(FWER) due to multiple assessments of efficacy. The method used to adjust
the sample size of a single stage design to account for multiplicity can
also be used to adjust the information level of an information monitored
design.
}
\examples{
# To be added
}
\seealso{
\code{\link[=required_information_fixed_n]{required_information_fixed_n()}} and
\code{\link[=required_information_mw_fixed_n]{required_information_mw_fixed_n()}} for obtaining
\code{information_fixed}; \code{\link[rpact:getDesignGroupSequential]{rpact::getDesignGroupSequential()}} for
obtaining \code{gsd_design_specification}.
}
