% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_by_id.R
\name{resample_by_id}
\alias{resample_by_id}
\title{Create resampled dataset from a list of resampled IDs}
\usage{
resample_by_id(data, ids_to_sample, convert_id_to_factor = TRUE)
}
\arguments{
\item{data}{A \code{data.frame} to be resampled, which must contain a column
named \code{.id} which uniquely identifies all rows corresponding to an
individual.}

\item{ids_to_sample}{a vector of IDs to be resampled from \code{data}}

\item{convert_id_to_factor}{Logical scalar: Should the ID column be converted
to a factor?}
}
\value{
A \code{data.frame} containing the resampled data with unique values
of \code{.id}.
}
\description{
This is an internal function used by \link{calculate_covariance} for
bootstrapping data that may be either in wide format (i.e. one row per
individual) or long format (i.e. one row per study visit per individual).
This function takes in a list of resampled IDs, retrieves all rows of data
corresponding to these IDs, and creates a unique ID that preserves the
nesting structure of long format of data.
}
\details{
The original data must have a column named \code{.id} that identifies which rows
of data belong to an individual. A new value of \code{.id} is created in the
resulting data, preserving the nesting structure in long data.
}
\examples{
# To be added

}
