% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_design.R
\name{rr_n_per_arm}
\alias{rr_n_per_arm}
\alias{rr_power}
\alias{rr_alpha}
\alias{rr_minimal}
\alias{rr_design}
\title{Calculate relative risk, sample size, power, or Type I error}
\usage{
rr_n_per_arm(
  pi_1 = NULL,
  pi_0 = NULL,
  rr = NULL,
  power = 0.8,
  alpha = 0.05,
  test_sides = 2
)

rr_power(
  pi_1 = NULL,
  pi_0 = NULL,
  rr = NULL,
  n_per_arm,
  alpha = 0.05,
  test_sides = 2
)

rr_alpha(
  pi_1 = NULL,
  pi_0 = NULL,
  rr = NULL,
  n_per_arm,
  power = 0.8,
  test_sides = 2
)

rr_minimal(
  n_per_arm = NULL,
  pi_0 = NULL,
  rr_above_1,
  power = 0.8,
  alpha = 0.05,
  test_sides = 2
)

rr_design(
  n_per_arm = NULL,
  pi_1 = NULL,
  pi_0 = NULL,
  rr = NULL,
  rr_above_1 = NULL,
  power = NULL,
  alpha = NULL,
  test_sides = 2
)
}
\arguments{
\item{pi_1}{A \code{numeric} vector: the risk of the outcome under treatment}

\item{pi_0}{A \code{numeric} vector: the risk of the outcome under control}

\item{rr}{A \code{numeric} vector: the relative risk of the outcome under
treatment relative to control}

\item{power}{A \code{numeric} vector: the statistical power of interest}

\item{alpha}{A \code{numeric} vector: Type I Error probability}

\item{test_sides}{A \code{numeric} vector: Number of sides for test (1 or 2)}

\item{n_per_arm}{A \code{numeric} vector: the sample size in each arm in a
1:1 randomized study}

\item{rr_above_1}{A \code{logical} scalar: search for a relative risk above
1 (\code{TRUE}) or below 1 (\code{FALSE})}
}
\value{
a \code{numeric} vector or \code{data.frame}, depending on the
arguments supplied and quantity of interest.
}
\description{
This is a function for designing a study with a binary outcome whose
inferential target is the relative risk. \code{rr_design()} will calculate
the minimum number sample size, relative risk, power, or Type I error when
the other arguments are specified. The approximation used relies on the delta
method for the log relative risk.
}
\details{
The function \code{rr_design} takes in user input, and depending on the
arguments supplied, calls the appropriate function depending on whether the
sample size (\code{rr_n_per_arm}), power (\code{rr_power}), Type I Error
(\code{rr_alpha}), or relative risk (\code{rr_minimal}),
}
\examples{
# Sample Size
rr_design(
  pi_0 = 0.08,
  rr = 0.8,
  power = 0.80,
  alpha = 0.025,
  test_sides = 1
)

# Power
rr_design(
  n_per_arm = 4500,
  pi_0 = 0.08,
  rr = 0.8,
  alpha = 0.025,
  test_sides = 1
)

# Type I Error
rr_design(
  n_per_arm = 4500,
  pi_0 = 0.08,
  rr = 0.8,
  power = 0.80,
  test_sides = 1
)

# Relative Risk: Treatment < Control
rr_design(
  n_per_arm = 4500,
  pi_0 = 0.08,
  rr_above_1 = FALSE,
  power = 0.80,
  alpha = 0.025,
  test_sides = 1
)

# Relative Risk: Treatment > Control
rr_design(
  n_per_arm = 4500,
  pi_0 = 0.08,
  rr_above_1 = TRUE,
  power = 0.80,
  alpha = 0.025,
  test_sides = 1
)
}
\seealso{
\code{\link[=hr_design]{hr_design()}} for hazard ratios.
}
