% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_colon_cancer}
\alias{sim_colon_cancer}
\title{sim_colon_cancer: Processed Moerton Colon Cancer Data}
\format{
\subsection{\code{sim_colon_cancer}}{

A tibble: 1030 × 18
\describe{
\item{.id}{Participant ID}
\item{arm}{Treatment Assignment}
\item{age}{Patient age}
\item{sex}{Patient sex}
\item{obstruction}{Colon obstruction by tumor}
\item{perforation}{Colon perforation by tumor}
\item{organ_adherence}{Organ adherence by tumor}
\item{positive_nodes}{Lymph nodes with detectable cancer}
\item{differentiation}{Differentiation of tumor}
\item{local_spread}{Extent of local spread of tumor}
\item{time_surgery_registration}{Time from surgery to registration}
\item{event_death}{Death on study (1 = yes, 0 = no)}
\item{event_recurrence}{Recurrence (1 = yes, 0 = no)}
\item{years_to_death}{Time to death/censoring in years}
\item{years_to_recurrence}{Time to recurrence/censoring in years}
\item{tx}{Binary Treatment Indicator (1 = LEV+5FU, 0 = LEV)}
\item{enroll_time}{Simulated recruitment time in years}
}
}
}
\usage{
sim_colon_cancer
}
\description{
Data from an adjuvant chemotherapy trial for colon cancer with variables
renamed and labeled, missing covariates imputed, and time scale converted
to years. For more information, see \code{?survival::colon}.
}
\keyword{datasets}
