% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speffsurv_impart.R
\name{speffsurv_impart}
\alias{speffsurv_impart}
\title{Wrapper for speff2trial::speffsurv: See \link[speff2trial:speffSurv]{speff2trial::speffSurv}}
\usage{
speffsurv_impart(
  data,
  estimand = "log_hazard_ratio",
  formula,
  treatment_column = NULL,
  alpha = 0.05,
  ci = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{estimand}{A \code{character} vector - must be "log_hazard_ratio" (default)}

\item{formula}{A \code{formula} for covariate adjustment.}

\item{treatment_column}{A \code{character} scalar, indicating the column containing the treatment indicator.}

\item{alpha}{A \code{numeric} scalar, indicating Type I error rate}

\item{ci}{A \code{logical} scalar: should CI be returned}
}
\value{
A \code{list} containing the estimate, standard error, estimand,
and confidence interval (if requested).
}
\description{
Wrapper for speff2trial::speffsurv: See \link[speff2trial:speffSurv]{speff2trial::speffSurv}
}
\examples{
speffsurv_impart(
  data = sim_colon_cancer,
  estimand = "log_hazard_ratio",
  formula =
    survival::Surv(time = years_to_death, event = event_death) ~
    age + sex + obstruction + perforation + organ_adherence + positive_nodes +
    differentiation + local_spread,
  treatment_column = "tx",
  alpha = 0.05,
  ci = FALSE
)
}
