% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_outcomes.R
\name{count_outcomes}
\alias{count_outcomes}
\title{Count outcome events from prepared study data}
\usage{
count_outcomes(prepared_data, study_time = NULL)
}
\arguments{
\item{prepared_data}{A prepeared dataset: see
\link{prepare_monitored_study_data}}

\item{study_time}{A \code{numeric} scalar indicating the study time at which
the data should be reconstructed and events should be counted}
}
\value{
A \code{data.frame} in long format, with one row per individual per
event (i.e. randomization, outcomes).
}
\description{
Count outcome events from prepared study data
}
\examples{
# To be added
}
