% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_to_n.R
\name{information_to_n}
\alias{information_to_n}
\alias{information_to_n_continuous_1_to_1}
\alias{information_to_n_binary_1_to_1}
\alias{information_to_events_log_hr}
\title{Convert information level into an approximate sample size requirement for a
one-stage design}
\usage{
information_to_n_continuous_1_to_1(
  information,
  sigma_0,
  sigma_1,
  round_up = TRUE
)

information_to_n_binary_1_to_1(
  information,
  pi_0 = NULL,
  pi_1 = NULL,
  delta = NULL,
  round_up = TRUE
)

information_to_events_log_hr(
  information,
  allocation_ratio = 1,
  round_up = TRUE
)
}
\arguments{
\item{information}{Numeric vector containing the information level}

\item{sigma_0}{Variance of outcomes in the population of individuals
receiving the control intervention}

\item{sigma_1}{Variance of outcomes in the population of individuals
receiving the active intervention}

\item{round_up}{Logical scalar: should the sample size be rounded up to an
integer value?}

\item{pi_0}{Probability of event in the population of individuals
receiving the control intervention}

\item{pi_1}{Probability of event in the population of individuals
receiving the control intervention}

\item{delta}{The risk difference (i.e. \code{pi_1 - pi_0})}

\item{allocation_ratio}{The allocation ratio of participants receiving
treatment to those receiving control}
}
\value{
When all parameters are scalars, the result is a scalar, indicating
the approximate sample size requirement. When multiple values are specified,
a grid of unique parameters are constructed, and the approximate information
is computed for each value of the parameters.
}
\description{
These functions provide an asymptotic approximation to the information
(i.e. precision, inverse of the variance) provided by two samples under
assumed values of nuisance parameters for continuous and binary outcomes.
}
\examples{
# To add

}
\seealso{
\code{\link[=asymptotic_information_difference_means]{asymptotic_information_difference_means()}},
\code{\link[=asymptotic_information_difference_proportions]{asymptotic_information_difference_proportions()}}, and
\code{\link[=asymptotic_information_mann_whitney_fm]{asymptotic_information_mann_whitney_fm()}} for an asymptotic
approximation of the information for a given sample size and values of the
nuisance parameters.
}
