% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_information_single_stage.R
\name{required_info}
\alias{required_info}
\alias{required_information_single_stage}
\alias{required_information_mw_single_stage}
\title{Determine the information level required for a one-stage, fixed sample size design}
\usage{
required_information_single_stage(
  delta,
  delta_0 = 0,
  alpha = 0.05,
  sides = 2,
  power = 0.8
)

required_information_mw_single_stage(mw, alpha = 0.05, sides = 2, power = 0.8)
}
\arguments{
\item{delta}{Numeric vector containing the estimand under the alternative hypothesis}

\item{delta_0}{Numeric vector containing the estimand under the null hypothesis}

\item{alpha}{Desired Type I Error Rate of the test}

\item{sides}{(Scalar - 1 or 2): Type of Test, either 1-sided or 2-sided.}

\item{power}{Desired power of the test (1 - Type II Error Rate)}

\item{mw}{Numeric vector containing the estimand under the alternative
hypothesis: the value of the Mann-Whitney under the null is always 0.5.}
}
\value{
When all parameters are scalars, the result is a scalar, indicating
the required information. When multiple values are specified, a grid of
unique parameters are constructed, and the required information is
computed for each value of the parameters.
}
\description{
In information monitored designs, data are accrued until the information
(i.e. precision) reaches the level required to achieve the desired power when
testing the null hypothesis against a specific alternative while maintaining
a specified Type I Error Rate.
\code{required_information_uninflated_mann_whitney} is an alias to
\code{required_information_single_stage} for ordinal outcomes, where the null
value of the Mann-Whitney estimand is 0.5.
}
\examples{
# When a single value is supplied for each parameter, a scalar is returned:
required_information_single_stage(
  delta = 5,
  delta_0 = 0,
  alpha = 0.05,
  sides = 2,
  power = 0.8
)

# When multiple values are supplied for one or more parameters, the grid of
# parameters are created, and a data.frame is returned.
required_information_single_stage(
  delta = c(5, 7.5),
  delta_0 = 0,
  alpha = 0.05,
  sides = 2,
  power = c(0.8, 0.9)
)
}
\seealso{
\code{\link[rpact:getDesignGroupSequential]{rpact::getDesignGroupSequential()}} for planning multi-stage
designs, and \code{\link[=required_information_sequential]{required_information_sequential()}} for adjusting the
information level from a single stage design to a multi-stage design.
}
